Autobuy moved to plugins, compile issues
