+++ project/.sonarlint/FLHook.slconfig
@@ -13,7 +13,7 @@
     },
     "C": {
       "ProfileKey": "AYDScucWudwity4Q9Znu",
-      "ProfileTimestamp": "2022-07-04T11:42:25Z"
+      "ProfileTimestamp": "2023-01-10T08:09:48Z"
     }
   }
 }
